pyproject.toml
[project]
name = "secure-ai-agent-matching-platform"
version = "0.1.0"
description = "Secure AI Agent Matching Platform with A2A Protocol"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]==1.19.0",
    "a2a-sdk>=0.3.16",
    "httpx>=0.28.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.25",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    "jsonschema>=4.23.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=23.2.1",
    "wandb>=0.18.0",
    "google-generativeai>=0.8.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
